refactor(replsession): extract evaluation pipeline from service
